refactor: clean up and organise
